requirements.txt
shapely
BeautifulSoup
Flask
elasticsearch